refactor: rename input component and update login form structure, import additional CSS files for styling
