+++ main/src/app/styles/input.css
@@ -1,0 +1,16 @@
+@layer utilities {
+  .border-input {
+    border-color: var(--input);
+  }
+  .bg-input {
+    border-color: var(--ring);
+  }
+
+  .focus-visible\:border-ring:focus-visible {
+    border-color: var(--ring);
+  }
+
+  .focus-visible\:ring-ring\/50:focus-visible {
+    box-shadow: 0 0 0 3px color-mix(in oklch, var(--ring) 50%, transparent);
+  }
+}
