+++ main/src/features/user/login-form/model/use-login.ts
@@ -1,16 +1,18 @@
 import { usePostLogin } from '@/shared/services/api';
 
+import { useRouter } from 'vue-router';
+
 type UseLoginReturn = {
   login: ReturnType<typeof usePostLogin>['mutate'];
   isPending: ReturnType<typeof usePostLogin>['isPending'];
 };
 
 export function useLogin(): UseLoginReturn {
-  // const router = useRouter();
+  const router = useRouter();
   const { mutate: login, isPending } = usePostLogin({
     mutation: {
       onSuccess: () => {
-        // router.push(dashboardPath.base.path);
+        router.push('/');
       },
     },
   });
